Add 'broken' status
If either a dkms module source, or the symbolic link pointing to
it is missing, the output of `dkms status` will be messed up.
Add a new status called 'broken', which will inform the user
about it in a nicely formatted way.
is_module_added() was modified to not report a module as added,
if not both the source directory and the 'source' symlink exist.
The man page was updated with the new 'broken' status.

Signed-off-by: Mart Frauenlob <AllKind@fastest.cc>